Package 'bin/llama_gemm' to wheel  (#320)
* pack llama_gemm

* update CMakeLists.txt

* remove candidate

* update MANIFEST.in
